[ fix ] rds 패키지 변경
테스트 코드 작성시 빈을 로드하는 context 지점을 최상위로 둘 수 있어야 하기 때문에
java.main.module 패키지 생성후 기존 패키지 모두 하위로 이동
