+++ module-rds-repo/src/main/java/module/entity/BaseEntity.java
@@ -1,4 +1,4 @@
-package entity;
+package module.entity;
 
 import java.time.LocalDateTime;
