+++ module-rds-repo/src/main/java/module/entity/Rating.java
@@ -1,13 +1,10 @@
-package entity;
+package module.entity;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
-import jakarta.persistence.FetchType;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
-import jakarta.persistence.JoinColumn;
-import jakarta.persistence.ManyToOne;
 import lombok.AllArgsConstructor;
 import lombok.NoArgsConstructor;
 import lombok.experimental.SuperBuilder;
@@ -16,20 +13,16 @@ import lombok.experimental.SuperBuilder;
 @SuperBuilder
 @NoArgsConstructor
 @AllArgsConstructor
-public class Seats extends BaseEntity{
+public class Rating extends BaseEntity{
 
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	@Column(name = "seat_id")
+	@Column(name = "rating_id", nullable = false)
 	private Long id;
 
-	@ManyToOne(fetch = FetchType.LAZY)
-	@JoinColumn(name = "screen_id", nullable = false)
-	private Screen screen;
+	@Column(nullable = false)
+	private String name;
 
 	@Column(nullable = false)
-	private String row;
-
-	@Column(nullable = false)
-	private Integer number;
+	private String img;
 }
