+++ module-rds-repo/src/main/java/module/repository/GenreRepository.java
@@ -1,8 +1,8 @@
-package repository;
+package module.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import entity.Genre;
+import module.entity.Genre;
 
 public interface GenreRepository extends JpaRepository<Genre,Long> {
 }
