+++ module-rds-repo/src/main/java/module/repository/MovieRepository.java
@@ -1,8 +1,8 @@
-package repository;
+package module.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import entity.Movie;
+import module.entity.Movie;
 
 public interface MovieRepository extends JpaRepository<Movie,Long> {
 }
