+++ module-rds-repo/src/main/java/module/repository/ScreenRepository.java
@@ -1,8 +1,8 @@
-package repository;
+package module.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import entity.Screen;
+import module.entity.Screen;
 
 public interface ScreenRepository extends JpaRepository<Screen,Long> {
 }
