+++ module-rds-repo/src/test/java/module/RdsTempContext.java
@@ -1,0 +1,8 @@
+package module;
+
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class RdsTempContext {
+	private void contextLoad(){}
+}
