+++ module-rds-repo/src/test/java/module/util/PrivateGetSet.java
@@ -1,4 +1,4 @@
-package util;
+package module.util;
 
 import java.lang.reflect.Field;
